finished refactor(I hope so)
