+++ serv/service/userService.go
@@ -3,66 +3,47 @@ package service
 import (
 	"stockpille/entity"
 	"stockpille/repository"
+	"stockpille/repository/entityRepository"
 	"stockpille/util"
 )
 
 func validateUser(user string, pass string) bool {
-	userDB := repository.GetUserByName(user)
+	userDB, err := entityRepository.Query(&entity.User{}, 2, user)
+
+	if err != nil {
+		return false
+	}
 
 	if userDB == nil {
 		return false
 	}
-	return util.ComparePass(userDB.Password, pass)
-}
 
-func NewUser(user entity.User) bool {
-	if repository.GetUserByName(user.Name) != nil {
-		return false
+	camps := userDB[0].GetCamps()
+	data := userDB[0].GetData()
+
+	for i := 0; i < len(camps); i++ {
+		if camps[i] == "password" {
+			passDB := data[i-1].(*string)
+			return util.ComparePass(*passDB, pass)
+		}
 	}
-	user.Password = util.Encript(user.Password)
-	repository.InsertUser(user)
 
-	return true
+	return false
 }
 
 func Login(user entity.User) string {
 	if validateUser(user.Name, user.Password) {
 		token := util.RandomToken()
-		repository.NewToken(token, repository.GetUserByName(user.Name).Id)
+		queryResult, err := entityRepository.Query(&user, 2, user.Name)
+
+		if err != nil {
+			print(err)
+			return "error"
+		}
+
+		repository.NewToken(token, queryResult[0].GetId())
 		return token
 	} else {
 		return "error"
 	}
-}
-
-func GetAllUsers() []entity.User {
-	return repository.GetUsers()
-}
-
-func GetUserById(id int) *entity.User {
-	return repository.GetUserById(id)
-}
-
-func DeleteUser(id int) {
-	repository.DeleteUser(id)
-}
-
-func UpdateUser(user entity.User) {
-
-	dbUser := repository.GetUserById(user.Id)
-	if dbUser == nil {
-		return
-	}
-
-	if user.Name != "" {
-		dbUser.Name = user.Name
-	}
-	if user.Role != "" {
-		dbUser.Role = user.Role
-	}
-	if user.Password != "" {
-		dbUser.Password = util.Encript(user.Password)
-	}
-
-	repository.UpdateUser(*dbUser)
 }
